Add VReg parameter to gen_{spill,reload,move}.
As described in Issue #14, this allows the client to make decisions
about the instruction to generate based on the virtual-register value
that is being spilled/reloaded/moved. This is useful when the register
class alone is more coarse-grained than we would like: e.g., if all
floating-point (F32 and F64) values use 128-bit vector registers (V128),
then we don't want to spill a V128 into a 16-byte spillslot if we're
only storing an F64.

Logically, one can think of this as dealing with (RReg, VReg) and
(SpillSlot, VReg) storage locations, rather than just RReg or SpillSlot:
we always talk about real storage together with the program value (VReg)
it is storing.