Add function liveins/liveouts to interface and backtracking allocator.
This change allows the Function to specify that certain registers
(probably real registers, though the interface is general) are live on
function entry -- i.e., have an implicit def before the function runs --
and/or are live on function exit -- i.e., have an implicit use after the
function returns. (Every block without successors is considered a return
point.)

This allows code generators that expect function arguments in certain
real registers, and return the function return value(s) in certain real
registers, to properly use the args and define the retvals without the
allocator complaining (undefined args) or miscompiling (spilled
retvals).

This change modifies the interface and updates the backtracking
allocator, but not the linear-scan allocator. The builtin tests do not
use this feature; I plan to test it from within Cranelift for now
(though we should add better tests here that exercise some sort of ABI,
too).