This implementation handles virtual-to-virtual, virtual-to-real and real-to-virtual coalescing.
The implementation has two parts.  The first is an analysis phase which
processes the output of |is_move| to find all possible opportunities.  This is
surprisingly complex, but it's a one-time piece of complexity.

The second component tries to use the collected information in the main
allocation loop.  At the moment it is very un-aggressive, and so is likely
missing many coalescing opportunities.  It will be further developed as I
understand more about the tradeoffs involved; but let it be said, as of this
commit, we are deep into the Land Of Heuristics.

Coalescing can be disabled trivially in the short term, if it causes problems,
by uncommenting a single line.