Rewrite the BT top level allocation logic, so that:
* There is no "searching, within each real reg assignment, for the cheapest
  eviction candidate that will allow some new VLR to be added for this rreg".
  This is a completely nonsensical concept (how did it ever get added?!) since
  the required set of evictees is unambiguous and can be calculated directly
  from the intersection of the VLR to be added and this register's existing
  commitments.  The top level is rewritten accordingly.  This improves
  performance a lot on large inputs.

* As a side effect of the above, the concept of a single evictee has been
  removed.  Instead BT will compute a set of evictees, and their total spill
  cost, needed to service an allocation request to a register.  This makes
  much more sense.

* Also as a side effect of the above, the main allocation loop now unifies the
  previously separate logic for "find a reg that can take this VLR without any
  eviction" and "find the reg that will take this VLR with the minimum evict
  cost".  The former is merely a special-case of the latter where the
  evict-set is empty.

Supporting changes:

Add a type `SpillCost`, rather than using `Option<f32>`, and some supporting
methods.  This makes it easier to do arithmetic and comparisons on spill
costs.

Remove almost all functionality associated with SortedRangeFragIxs, reducing
its role to a simple container for FragIxs, primarily used to store the
fragment indices for each RealRange and VirtualRange.

In the BT allocator, the per-real-reg commitment maps have been abstractified
into their own type, `CommitmentMap`.  This is currently implemented using an
enhanced version of the existing `SortedRangeFragIxs`, but moving it into its
own type makes it possible to replace the implementation with something
better, probably a balanced interval tree or something using bitmaps.

This commit introduces the concept of having a simple, slow, reference
implementation of computing interferences, `find_Evict_Set_CROSSCHECK`, which
can be used to debug the production implementation `find_Evict_Set_FAST`.  The
latter needs to be fast and correct, and that's difficult; having a
cross-check implementation proved to be a great help.