A further improvement (well, maybe improvement) to coalescing in the BT allocator, aimed at giving more consistent coalescing for large VLR groups.
With this patch, a VLR can have "secondary hint" registers, which are found by
looking at the assignments for all other VLRs in the group.  This means that a
VLR can "request" the same reg that has been assigned to a VLR in the group
from which it is "distant", that is, it has no direct copy-insns that connect
it to the other VLR, and instead is only transitively connected to it.  Prior
to this commit a VLR preference could only be acquired from an immediately
neighbouring VLR in the group.

I suspect this is about as aggressive as I can make coalescing be, without
rewriting the core logic to allocate entire groups of VLRs at once, as
described in the TODO section in bin/minira.rs.