Use the new AVL trees to re-implement CommitmentMap.  This produces allocations much more quickly for large functions.  It's a big-O improvement, not just a constant-factor speedup.
A very slow cross-check implementation is retained for debugging/verification
until such time as we have more confidence in this new implementation.  The
cross-check implementation is disabled in this commit.  To enable it, change
CROSSCHECK_CM at the top of backtracking.rs to |true|.  Both implementations
will then be run in parallel and be crosschecked frequently.