Backtracking allocator: change VirtualRangePrioQ's implementation from the stand-in linear-array implementation to a production-grade binary-heap based implementation.  This gives significant speedups on very large inputs (functions).
Note that this may change the resulting allocations, because it sometimes
changes the order in which virtual live ranges are processed.  Even in the
worst case the effect is small and only appears to affect the situation where
there is a lot of spilling.  The resulting allocations will of course still be
correct.