Short term (but large) speed improvements in computation of dominator sets. Due to me missing a tiny footnote in Muchnick's description of the simple iterative algorithm, my implementation of it was, I suspect, O(N^3) where N is the number of nodes.  It should have been O(N^2).  Changing the node iteration order from "arbitrary" to reverse postorder achieves the O(N^2) time bound.  I must thank Chris Fallin for indirectly bringing this to my attention.
Note that this is an interim fix that significantly reduces run time for some
of the large test cases we have.  The longer term plan is to switch to the
very much faster scheme presented in "A Simple, Fast Dominance Algorithm
(Keith D. Cooper, Timothy J. Harvey, and Ken Kennedy)" which should give
production-grade performance and may allow us to avoid the overhead of
computing dominance sets for each node, by working instead with a single
dominance tree for the entire function.  This is yet to be assessed.

This commit also adds some more structuring, assertions and debug printing to
CFG::create().