Bugfix in checker dataflow analysis: don't meet inputs of all succs.
Previously, the dataflow analysis loop would take the state after
passing through all insts' transfer functions and meet it with current
state at each successor BB's input to update that successor. However,
the previous code erroneously reused this state, so in effect it would
also meet the input of a second successor with that of the first. This
would lead to a correct but nonminimal dataflow solution, which in turn
would lead to false positives.

Consider the particular testcase `fuzz_checker1` found by Julian Seward:

```
v26I = i32
b0:
  imm     v26I, 1499027801
  if_then_else v26I, b2:b2, b2:b2
b1:
  if_then_else v26I, b2:b2, b2:b2
b2:
  if_then_else v26I, b0:b0, b1:b1
```

which can be reproduced with this command line:

```
$ cargo run -- -a btc -t fuzz_checker1 -f 8 -i 8
```

At the input to `b0`, `r0` has state `Conflicted`, because it is the
meet of the symbolic live-in realreg `r0` and the def of `v26` carried
across the `b2 -> b0` edge. This is fine, because the `imm` instruction
overwrites this value in `r0`, and this def reaches all real uses.
However, this `Conflicted` value was also erroneously mixed into the
input state of `b1`, from which it also reached the input state of `b2`.
This patch fixes the above test-case (it passes the checker now).