Change the representation of registers harvested from the client's |get_regs| routine.  Before, three |Set<Reg>|s were created for each instruction.  Now, info for all instructions in a function is placed in just four large vectors, contained in a |RegVecsAndBounds| structure.  This new interface to register-use information has been connected to its various use points.  This reduces the allocator's instruction count by around 10% and the number of heap blocks allocated by around 25%.
In the external interface, the type of |get_regs| has changed.  It is now
passed a |&mut RegUsageCollector|, an opaque object owned by the allocator.
All it has to do is call |add_use|, |add_def| or |add_mod| on the object,
passing it the relevant |Reg|.

The semantics surrounding sanitisation of register uses, and how this
interacts with the 'proposed scratch register', has been extensively
re-checked.  From the interface perspective it is unchanged, although the
interface comments explaining it have been clarified.  Internally, there is a
small change: in the incoming vcode, write-uses of the reserved registers were
previously removed at sanitisation time for LSRA but not for BT.  Now they are
not removed for either.  This does not appear to be a problem for LSRA -- all
such writes are dead.  This simplification means that the |sanitize_scratch|
flag passed to sanitization is no longer needed.