Add `orig_insn_map`: mapping from new to original insn indices.
This is needed in the Cranelift client code in order to track
source-location mapping for debug info: we need to be able to shuffle
the source locations into the new locations after the regalloc has done
its instruction-stream editing. The `target_map` result is not quite
good enough, because it only provides old --> new mappings at a basic
block granularity.

This commit also adds "invalid" values to the index types; this is
useful to represent "new instruction maps to no old instruction" without
the memory-layout waste of an `Option<u32>`.