Use incrementally-updated, dense vreg-to-rreg maps.
This PR changes the interface of `map_regs()` to take, instead of two
Maps (a hashmap), an abstracted `RegUsageMapper`.
This data structure represents the "before" and "after" (use-point and
def-point) maps as, respectively, a dense vector-based map and a sparse
delta overlay, with the latter sorted if the delta is large.
The dense array-based map should be faster in almost all cases than the
hashmap, and the delta overlay should minimize memory usage (as compared
to maintaining two separate maps).

This PR also changes the instruction-stream editing to maintain running
incrementally-updated maps using this data structure, rather than
cloning the full hashmap twice for every instruction. A preliminary
investigation of Cranelift's new backend using DHAT shows that this
clone is very costly in terms of allocation. This interface and
algorithm change, when adopted in Cranelift, seem to reduce total
allocated bytes during a Wasm module compilation by about 15%.