For the backtracking allocator: an improvement to coalescing.  When creating spill/reload live ranges, see if we can also create a coalescing hint for the range.  This is motivated as follows (using arm64 as an example): imagine the vcode contains a vreg-to-vreg move (very common):
   mov %vDst, %vSrc

which is identified as a virtual-to-virtual boundary move -- meaning, one that
can be coalesced.  If one of the operands, say %vSrc is spilled but the other
(%vDst) is assigned a real register (%rDst), then we wind up with the
post-allocation result having a redundant copy:

   reload %rTmp, [spill-slot]
   mov %rDst, %rTmp

This happens because the spilled range (%vSrc) is assigned its own register
(%rTmp) but the allocator is unaware that we'd really like it to be in the
same register as whatever %vDst gets (or got) allocated to.

This commit fixes that, by attaching a coalescing hint to %vSrc when the
decision is made to spill it, saying that it would prefer to ride in the same
register as %vDst does (or will).  Then, later, when the main allocation loop
comes to allocate %vSrc, the existing pay-attention-to-coalescing-hints
machinery will try to give it the same register as %vDst.  Hence we get simply

   reload %rDst, [spill-slot]

which is what we want.  The same technique is applied the other way around if
%vDst is spilled and %vDst gets a register.

Note also, the above explanation is written as if there is no difference
between virtual registers and virtual live ranges.  In reality there is a
difference, and we're allocating virtual live ranges, not virtual registers.
But the explanation remains correct.