For the backtracking allocator: a second improvement to coalescing.
This change improves accuracy of spill cost estimation in the presence of
coalescing, in the hope (which sometimes comes true) of reducing the total
amount of spilling.  It is motivated by the following observation:

Imagine we have two virtual live ranges (VLRs) (for simplicity here, just
think of them as virtual registers): %v1 and %v2, used like this:

   %v1 = ...     // a def of %v1 (in fact the only def)
   ...
   %v2 = %v1
   ...
   ... %v2 ...   // a use of %v2 (in fact the only use)

In the analysis phase, the allocator computes a total use count for each VLR,
based on the number of "mentions" of it in each block, scaled by the estimated
execution count for the block.  This is then used to compute a spill cost for
the VLR, which in turn is use to determine its priority when allocating
registers to VLRs.  That sounds reasonable, and mostly it is.

However, in the presence of coalescing, these counts become inaccurate.  Prior
to this patch, both %v1 and %v2 would be regarded as each having two
"mentions", and their spill costs would be calculated accordingly.  However,
if %v1 and %v2 are coalesced, as we hope, then the move instruction
disappears, which means that each range now only has one mention.  This should
mean that its priority for getting a register should be reduced accordingly.
However, the current code does not do that.

In an extreme case, a VLR would be coalesced with all its "neighbours" and not
have any uses.  Hence it logically would simply disappear; yet the fact that
its spill cost remains non-zero means it will eventually wind up competing
pointlessly for registers against other VLRs that really do have a use for a
register.

This patch tries to correct that during coalescing analysis.  It adjusts a
VLR's spill costs downwards if it is a coalescing candidate, and can even
adjust it down to zero.  Note that this "fix" is still inaccurate in that it
only takes into account whether a VLR is a coalescing candidate, but it
ignores the reality of whether or not coalescing actually happened.  These
heuristics could all do with a re-think and re-work at some point.

The effect is mostly a small improvement in benchmark insn counts (meaning an
overall reduction spilling).  Overall it appears to be a win in terms of
generated code quality.  Interestingly, reductions in allocation time of up to
25% have also been observed.