linear scan: A Revamp;
This changes a lot of things:

- adds statistics for counting some interesting events,
- simplifies the representation of intervals, based on the fact that
inactive intervals can only be fixed (real) registers intervals.
Everything is explicitly taken by ownership, which makes things simpler.
- disentangle a few allocator responsabilities:
    - updating state is done as part of the ActivityTracker, which is
    also simpler now. Indeed, only the fixed (real) registers intervals
    can be inactive, which remove a bunch of cases, and
    - unhandled intervals are separated in a small data structure using
    the AVL tree for efficient insertion/removal in random indexes.
- keep track of vreg mentions in instructions, and split those along
with intervals, then reuse this information to implement a simplistic
yet efficient RegUsageMapper. Right now this is done when constructing
live intervals, which is inefficient and could be avoided by doing it
during analysis (later patch).
- remove inefficient algorithms that were only used in debug-only mode
to compare with newer implementations of equivalent algorithms.