BT: clean up and document some interactions between the coalescing analysis and the spill slot allocator.
The coalescing analysis computes, amongst other things, equivalence classes of
VirtualRanges.  Two VirtualRanges will be assigned the same equivalence class
if there is a move instruction that transfers a value from one range to the
other.  The equivalence classes created are the transitive closure of this
pairwise relation.

The main allocation loop tries to assign the same register to all the
VirtualRanges in an equivalence class.  Similarly, the spill slot allocator
tries to allocate the same spill slot to all the VirtualRanges in an
equivalence class.  In most cases they are successful, and so the moves
between those VirtualRanges will later disappear.  However, the complete story
is not quite so simple.

It is only safe to assign the VirtualRanges in the same equivalence class to a
single register or spill slot if those VirtualRanges are non-overlapping.
That is, if their overall collection of RangeFrags is disjoint.  If two such
VirtualRanges overlapped, then they could be carrying different values, and so
they would need separate registers or spill slots.

Most of the time, these equivalence classes are indeed internally
non-overlapping.  But that's just luck -- that's how the input VCode mostly
is.  The coalescing analysis *doesn't* check for overlaps within an
equivalence class, so it can be the case that the members of an equivalence
class overlap.  The users of that information -- the main allocation loop and
the spill slot allocator -- currently check for, and handle, such situations.
So the generated allocation is correct.

It does, however, cause imprecision and unnecessary spilling, and, in the main
allocation loop, slightly increased evictions.  Worse, though, is that these
invariants are all undocumented.

This patch makes the following changes:

* It documents them; the above comment (more or less) is added at the start of bt_coalescing_analysis.rs.

* In bt_coalescing_analysis.rs, there is a half-hearted effort made to avoid
  creating equivalence classes whose elements (VirtualRanges) overlap.  This
  is done by doing an overlap check on two VirtualRanges connected by a move,
  and not merging their equivalence classes if they overlap.  That helps, but
  it doesn't completely avoid the problem because there might be overlaps
  between other members (VirtualRanges) of the about-to-be-merged equivalence
  classes.

  Fixing this properly would entail doing a complete overlap check for the two
  equivalence classes before merging.  That sounds expensive, though.  Hence
  this partial fix.

* In the spill slot allocator (bt_spillslot_allocator.rs), the search space
  (meaning, the available spill slots) checked in each call to
  `alloc_spill_slots` has been greatly reduced.  Now, only the
  highest-numbered 8 slots are searched for spaces.  That has two effects:

  - it avoids lots of pointless repeated checking of low-numbered spill slots,
    that long ago became full(ish) and are unlikely to be able to take any new
    VirtualRanges

  - more subtly, it interacts with the question of whether or not each
    VirtualRange equivalence class is internally overlapping.  When no
    overlaps are present, the spill slot allocator guarantees to find a slot
    which is free for the entire equivalence class, which is the ideal
    solution.  When there are overlaps present, the allocator is forced to
    allocate at least some of the VirtualRanges in the class to different
    slots.  By restricting the number of slots it can choose to 8 (+ extras if
    it needs them), we reduce the tendency for the VirtualRanges to be
    assigned a large number of different slots, which in turn reduces the
    amount of spilling in the end.

Of course the "proper" fix for all this is to fix the coalescing analysis so
as only to build non-internally-overlapping VirtualRange equivalence classes.

* Minor change: fix some misleading naming (vlrXX vs vlrixXX) in
  `do_coalescing_analysis`.

* Minor change: a bunch of commented out debug printing has been removed.

Some test cases take longer to compile, some less, but in no case is the insn
count difference greater than 0.1%.

For `joey_big.clif`, which spills heavily, static instruction counts are:

  BEFORE: 8151 spills, 8829 reloads, 337 total spill slots,
          69280 moves coalesced out at the spill level

  AFTER:  6477 spills, 7157 reloads, 356 total spill slots,
          70948 moves coalesced out at the spill level

so a significant reduction in static spill/reload counts, and only a modest
increase in the number of spill slots used.