BT: avoid construction of "no-evict" sets
When BT searches for a usable register to hold a VLR that belongs to an
equivalence class of VLRs (as determined by the coalescing analysis), it is
important not to choose for eviction, any other member of that class.  That
would undermine the whole point of coalescing.  It wouldn't lead to incorrect
allocations, though.

Until now, BT constructed, for each VLR, a SparseSet holding all the VLRs in
the same equivalence class, and passed it to `find_evict_set`.
`find_evict_set` then checks this set and will not request eviction of any
member in it.  The sets are constructed by consulting the
`UnionFindEquivClasses` created by coalescing analysis.

This works, but is wasteful: if there is a large equivalence class (of VLRs)
containing N elements, then the same no-evict set, of size N, will be
constructed N times.  This is quadratically bad in the worst case.

This patch removes all of the set construction and testing.  Instead,
`UnionFindEquivClasses` is enhanced so that it can directly answer the
question "are these two items in the same equivalence class?"  Then, a
suitable closure that calls the new function is passed to `find_evict_set`,
and that uses it to make the same evict/no-evict decision that it previously
did via set-membership tests of the no-evict set.

A closely related ridealong fix, is that all of the methods of `struct
SpillCost` are now marked as inline-always.  They were called a lot from
`find_evict_set` and so the inlining is a significant win.

Overall, for huge functions (`joey_big.clif`) this reduces insn count by 6.4%
and the number of malloc calls by 25.4% (!).