linear scan: Use a smallvec for Mentions;
Observing a distribution of the number of mentions shows that in our
test cases, a given virtual Interval ends up having one or two explicit
mention sets, at the end of register allocation (in particular, after
splitting).

This is a 10% improvement in terms of compile-time for
joey_{small/medium}, and a small statistically-not-significant
regression (<0.1%) on joey_big.