Checker: support coalesced spillslots.
The key idea of this change is to use "ghost instructions" to
communicate information from the register allocator to the checker.
These instructions have no side-effect, by definition, but communicate a
semantic change to the program state that is relevant to verification:
for example, a move from one virtual register to another, when both
virtual registers live in the same spillslot.

These ghost instructions exist only as edit-list entries; they are never
actually reified into real instructions and inserted, nor does the
machine backend have to have any special support to generate them. They
are just conceptual.

It turns out that to support spill-slot coalescing (i.e. to successfully
check allocation results with this feature enabled), all we need is to
correctly handle these V-V moves. Other forms of coalescing (e.g., a
move between two virtual registers allocated to the same real register)
are already visible to the checker, because it sees these redundant
moves before they are removed.

In a little more detail, the design-space considered included:

1. Extending the checker to reason about storing multiple virtual
   registers in each storage slot / allocation target (real register or
   spill slot). The problem with this approach is that it is much less
   efficient -- analysis values for each slot at each program point are
   no longer fixed 64-bit things but potentially unbounded sets -- and
   that it also loses information, because we may have two different
   paths that store a different vreg in a slot when arriving at a
   program point via different paths. Fundamentally, we should check the
   move at the point that the sets are union'd (the "ghost instruction")
   and then reason only about the (single) "current" vreg in the slot.

2. Reason about collections of equality constraints by assigning
   symbolic names (tags or whatever) to each def (i.e., SSA values) and
   then gathering move-constraints together. This started to get really
   complex, and seems to require basically a full SSA recovery algorithm
   plus a good part of the allocator's analysis, so it was discarded
   (the checker should be simple and straightforward to maximize the
   confidence we have in it).

3. Reason about reaching definitions, and validate the program by
   showing that the same defs reach each use at each instruction in the
   pre-allocated and post-allocated versions. The problem with this is
   similar to option 1 above: it discards information about paths, and
   so is imprecise.

Fundamentally, it seemed that we could either try to design clever
heuristics to *recover* the information about the elided V-V moves, or
we could just observe them directly before they are elided; so the above
ghost-instruction approach was chosen.