Add stackmap checking to the checker, and fix some stackmap fuzzing issues. With these changes, the `bt` fuzz target (with reftypes / stackmaps) reached 10.9B successful test cases.
The checker for stackmaps works as follows:

- Every analysis value in the checker gets a "reffy" bit. These are
  filled in according to the list of reffy vregs provided by the
  regalloc.

- When stackmap information is given for a safepoint, the checker
  verifies that every slot described as reffy by the stackmap is in fact
  reffy.

- Any reffy spillslot that passes through a safepoint *without* being
  named in the stackmap is set to "undefined". Then, if it is later used
  as a reference (or at all), we will detect that it was in fact live
  but was not named in the stackmap.

Note that these rules do *not* do the obvious thing and construct an
expected stackmap from the spillslot state at the safepoint's program
point, then compare it to the given stackmap. This is because a stackmap
is allowed to elide a reffy spillslot that's dead: we could just as
easily have zeroed it out after its last use, as it semantically no
longer exists in the program; we were just too lazy to clear it. We
don't want to reconstruct liveness information and verify that every
reffy-but-not-described-as-such slot is dead in the checker, so instead
we just mark every reffy slot not named as reffy as undefined. This is
somewhat intuitive when seen a certain way: if the GC (or whatever
stackmap consumer) is unaware of a reference value at a safepoint, it
might be outdated (e.g. if the GC moves every object), so might as well
be undefined.

The stackmap checking appears to truly work because it found the above
dead-spillslot issue quite quickly. After fixing that bug, it ran up to
the following state (10.9B test cases) before I killed it:

`#10943745210: cov: 4362 ft: 19953 corp: 2941 exec/s 1577 oom/timeout/crash: 0/0/0 time: 303772s job: 24310 dft_time: 0`