Fix slowness in BT's coalescing analysis when some regs have very many `RangeFrag`s.
This fixes extreme slowness in all the test cases attached to
https://github.com/bytecodealliance/wasmtime/issues/911, when compiling via CL's new x64
backend.

The fix involves replacing a linear search with a binary search, but only in certain cases.  The
implementation and supporting heuristics are somewhat involved.  The patch contains a large
block comment describing the problem and its solution in considerable detail, so those details
are not repeated here.  Suffice it to say that fixing the problem without impacting allocation
time for "normal" inputs required some care.

For one of the pathological test cases in PR911, file2.wasm.1, compilation time was reduced by a
factor of 56 x, from about 62 seconds down to 1.10 seconds (Intel Xeon, 2.6 GHz, single
compilation thread).

For more "normal" cases, the fix is perf (compiler insn count) neutral to a 6% win, but there is
an increase in allocated blocks of up to 0.7%.  This seems difficult to avoid.