Rewrite `AVLTree::insert_wrk` to be non-recursive.  Profiling shows this function to be hot in the case where the function to be allocated contains very many function calls, since then there will be a huge number of minimum-length real reg live ranges to be inserted into commitment trees.
The rewrite here was created by mechanical transformation of the recursive
original through many intermediate steps, all of which functioned correctly.

The supporting functions `leftgrown` and `rightgrown` have been split into hot
and cold parts for better inlining.

For the `joey_big` test case, this reduces compiler insn count by 1.1%;
smaller functions get smaller wins.  For some of the pathological test cases
in Wasmtime PR911, there is a wallclock time win of up to 2%.

From those pathological test cases it is also clear that, at least for
themselves, there is a much larger win on the table here: since the tens of
thousands of RangeFrags to be inserted into each tree are in order, it's
wasteful to insert them one by one.  Instead the sorted vector of them, that
is pushed into the tree, could be transformed in linear time into an almost
perfectly balanced binary tree.  Whether this is worth it for any realistic
inputs seems doubtful, however.