Import PrettyPrint support from Cranelift;
This imports the PrettyPrint primitives from Cranelift, and splits the
trait into two, so that there's no need to have a dynamic panic for the
implementation of the sized variant for Regs.

This also makes it possible to not expose WritableBase, since it was the
only use outside the regalloc crate in Cranelift. Yay for static typing!