Add `is_included_in_clobbers` to public API.
This allows the client to request that certain instructions' register
writes be excluded from the returned clobbered-register set. This may be
useful when the client can handle such writes in other ways: for
example, by reasoning directly about ABIs. (One common case is that if a
caller and callee have the same ABI, then the call instruction in the
caller need not be included in the clobber set that is used to generate
the caller's clobber-saves, because any register that the callee
clobbers is also allowed to be clobbered by the caller.)