`get_range_frags_for_block`: improve cost handling for 'modifies' references to a reg.
One of the tasks of `fn get_range_frags_for_block` is to count the number of times each virtual
reg is mentioned, so as to give a basis for computation of spill costs.  It adds one for a
'use' or 'def' reference, which is reasonable.  But it also currently adds one for a 'mod'
reference, which is a bit misleading, given that if we spill a live range containing a 'mod'
reference, we'll have to generate a reload before the referencing insn and a spill afterwards.
So really we should increment the mentioned-count by two, not one.  This patch makes that change.