Tweak the crate/workspace setup so that fuzz binaries are built in root `target/` directory.
Because the fuzzing crate was previously in its own Cargo workspace, its
build process placed its binaries in `fuzz/target/...`. Other
infrastructure (such as OSS-Fuzz) doesn't expect this. This PR instead
makes the fuzzing crate an ordinary part of the main workspace, and
tweaks the build-option overrides to work with this new setup. Followup
to #106.

Also re-runs `rustfmt` as its output has changed slightly with the new
Rust release.