Linear scan: add new invariant to allow placement of block fixup moves.
Linear scan requires some "space" to insert moves across blocks, just
before the last instruction in the block. This was assuming the block
terminator would have no register mentions (use, mod, def), which is not
generally the case: in particular, a branch table may use a register.

In this case, we can't put the block fixup moves before the block
terminator, and we have to put them in the successors. This is fine, as
long as the successors only have one predecessor; otherwise it would
mess with the control flow coming from the other predecessors.

This is marked in store as a new invariant for linear scan:
- either a block ends with a control flow instruction that has no
  register mentions,
- or the successors of the given block have only one predecessor.