Linear scan: precisely keep track of blocks start and end for each interval.
In the phase inserting the fix up moves across blocks, a particular edge
case could cause confusion. When two intervals pertained to the same
virtual register, and one "contained" (with respect to its bounds) the other,
there was no way to know which location was the canonical one for the
given vreg (since an interval is a vreg and all the related redefinitions/uses).

As a matter of fact, this means that each interval must precisely track
which blocks start/end it contains (and that must be maintained when
splitting the intervals). This is required for correctness, but also
opens the door to some future improvements (e.g. selection of split
positions).