Arena allocation using bumpalo (and hashbrown with custom alloc).
This PR modifies both linear scan and BT allocators (and their shared
analyses) to pervasively use `bumpalo` for memory allocation from an
arena that lives for the duration of a function's register allocation.
This should result in significant performance improvements (I have not
yet measured, though), as we have seen malloc/free to be significant
parts of CPU-time profiles in the past.

Care must be taken to *not* use normal heap-allocated types inside of a
`BumpVec`, `BumpMap` or `BumpSet`, because the arena does not invoke
`Drop` impls when it is finally discarded (this is what allows freeing
the memory to be so fast; we just discard the entire block at once). As
such, many uses of `SmallVec` and some other types had to change. Some
cases that I had missed were picked up by LeakSanitizer in fuzzing runs,
and these errors no longer occur, so I'm relatively satisfied that we've
covered them all.

This changes the public interface -- the caller must provide the arena,
and some of the returned metadata is allocated within it -- so it will
require corresponding updates in e.g. Cranelift. That will come in
subsequent PRs.

Thanks to @bnjbvr for noting in our discussion this morning that memory
allocation is still significant in profiles and thus inspiring me to
dive into this :-)