Add support for using regalloc2 as a backend register allocation algorithm.
This PR adds a shim that translates regalloc.rs' view of the input
program into a form that regalloc2 can use, and translates the
allocations back to regalloc.rs-like results.

This was co-developed and -optimized with regalloc2, as a sort of
short-to-medium-term solution to the Cranelift use-case before Cranelift
can be migrated to use the regalloc2 API directly. Performance-testing
results show that this is a marginal improvement over existing (current
default) BTRA, including the shim overhead, so I believe that this is a
reasonable thing to use as a stepping-stone rather than wait for the
above transition.

It also includes the fuzz target `regalloc2` that runs the regalloc.rs
checker with the new algorithm; I have fuzzed the combination of the two
with this target, in addition to fuzzing just regalloc2 with its own
fuzz targets.