Speed up the dominance check in calc_loop_depths
The current setup requires walking the whole dominance tree from a node
M to the root when checking if a node N dominates M. This process can
actually be stopped as soon as a node with post-order index greater than
that of N is reached. This can prevent accidentally quadratic behavior
in situations where there are chains of sequential blocks.